+++ readBvecsFromTwix.m
@@ -12,8 +12,15 @@ function readBvecsFromTwix(infile, outfolder, T)
 %   respectively. It will also output the nominal *b*-values to 
 %   `output_folder/input_twix_file_nominal.bval`.
 %
+%   You can alternatively provide an already generated twix object, for example:
+%     twixobj = mapVBVD('input_twix_folder/input_twix_file.dat')
+%     readBvecsFromTwix(twixobj, 'output_folder/');
+%   which will do the same as the first example. This can be useful to avoid
+%   having to repeat reading in the twix object if it is already being used
+%   elsewhere in a script.
+%
 %     readBvecsFromTwix('input_twix_folder/input_twix_file.dat', 'output_folder/', diag([1,-1,1]));
-%   will do the same as the previous example, but apply the transformation matrix
+%   will do the same as the first example, but apply the transformation matrix
 %   diag([1,-1,1]) to the b-vectors before saving them. This is useful to 
 %   e.g. apply the transformation from DICOM to NIfTI space.
 %
@@ -27,11 +34,33 @@ if ~exist('T','var')
 end
 
 %% read in data
-dat=mapVBVD(infile);
+% Option to provide filename or previously generated twix object
+if (isstring(infile) || ischar(infile)) && isfile(infile)
+    dat=mapVBVD(infile);
+elseif isstruct(infile) || (iscell(infile) && isstruct(infile{end}))
+    dat=infile;
+else
+    error('input "infile" does not seem to be either a filename or a twix object')
+end
 
-nRep=dat.image.NRep;
+% Twix files can contain multiple scans
+if iscell(dat)
+    if numel(dat)>1
+        warning('twix object contains more than one scan; using the last one')
+    end
+    dat=dat{end};
+end
+
+% Validate input
+assert(isstruct(dat),'input did not give twix struct')
+assert(isfield(dat,'image'),'input did not contain image data')
+assert(isa(dat.image,'twix_map_obj'),'input did not give valid twix object')
+
+% Basename for output files
+[~,basename,~]=fileparts(dat.image.filename);
 
 % Choose one line from each repetition
+nRep=dat.image.NRep;
 lastIndices=zeros(nRep,1);
 for n = 1:nRep
     lastIndices(n)=find(dat.image.Rep == n,1,'last');
@@ -59,8 +88,6 @@ bNominal=iceParams(7,:) -k;
 [bVectors,bValues]=readBvecsFromBmatrix(B,bNominal);
 
 %% output bvals and bvecs
-[~,basename,~]=fileparts(infile);
-
 saveBvec(bVectors,outfolder,basename,T);
 saveBval(bValues,outfolder,basename);