+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog uses [semantic versioning](https://semver.org/)
 ## [unreleased]
 - Breaking change: `readBvecsFromTwix` now only takes one file argument.
 Replace calls using the previous format `readBvecsFromTwix(infolder,infile,outfolder)` with `readBvecsFromTwix(fullfile(infolder,infile),outfolder)`
+- Option to provide twix object from mapVBVD rather than a filename to `readBvecsFromTwix`
 
 ## [v0.4.0]
 Added mapVBVD as a submodule